Fix DarkPanel; Fix PolyGnome rounding; Fix warning
1. DarkPanel support had broken us at head because I forgot
   a widget:;step
2. But also was unfinished since I hadn't actually removed the
   json read etc
3. While at it, fix a rounding error in seven segment display
   which closes #102
4. And fix a couple of warnings the docker toolchain shows

I'll mark this 2.4.0

+++ src/KSSynth.hpp
@@ -122,7 +122,7 @@ class KSSynth
         {
             for (int i = 0; i < burstLen; ++i)
             {
-                delay[i] = (float)rand() * 1.0 / RAND_MAX;
+                delay[i] = (float)rand() * 1.0 / (float)RAND_MAX;
                 delay[i] = delay[i] * 2.0 - 1.0;
             }
             break;
@@ -150,7 +150,7 @@ class KSSynth
             for (int i = 0; i < burstLen; ++i)
             {
                 delay[i] = (i * 1.0f / burstLen) - 0.5;
-                delay[i] += (float)rand() * 1.0f / RAND_MAX - 0.5;
+                delay[i] += (float)rand() * 1.0f / (float)RAND_MAX - 0.5;
             }
             break;
         }
